// src/app/models/usersDetails.model.ts
// user.model.ts
export class UserDetails {
    aboutYourSelf: any;
    address: any;
    city: any;
    country: any;
    description: any;
    email: any;
    fileData: any;
    fileName: any;
    fileType: any;
    fullName: any;
    phoneNumber: any;
    state: any;
    userId: any;
    zipCode: any;
  
  }
  